Expose the encoding id as its own u16 (#28)
Base zenoh's `EncodingId` is a `u16`: `Encoding::id() -> EncodingId` and
`Encoding::new(id: EncodingId, ..)`. flat widened it to `i32` on the way
out and narrowed it back with `as u16` on the way in, which was wrong
twice over.

`i32` is a binding-specific type — a signed 32-bit integer is one
target's integer width, and §Be faithful to zenoh forbids naming a wire
type in flat's surface. Choosing that lowering is the adapter's job.

The narrowing was a wrong-value bug with no diagnostic: an out-of-range
id silently became a *different, valid* encoding. `65536` arrived as `0`
and `-1` as `65535`, on a type that appears in every sample.

Both go away by using zenoh's own type:

- `encoding_get_id(&Encoding) -> u16`, returning `e.id()` uncast
- `EncodingStruct.id: u16`
- `encoding_new_from_id(id: u16, ..)`, passed to `Encoding::new`
  unchanged

There is no validation to add and no `Result` to introduce: every `u16`
is a valid encoding id, so the invalid inputs are now unwritable rather
than rejected at runtime.

`every_id_round_trips_losslessly` pins the accessor and the value form
across the domain and at both ends of it, including a schema-bearing
variant to show the schema does not perturb the id. Mutation-tested:
narrowing the accessor (`e.id() as u8 as u16`) or the constructor
(`id & 0xFF`) fails it at id 256. The pre-existing `id_round_trips` and
the wire-id assertions in `constants_render_canonically` needed no edit —
their literals infer as `u16`.

Closes #28

+++ src/base/bytes/encoding.rs
@@ -4,8 +4,8 @@ use crate::Encoding;
 
 /// Return the numeric identifier of an encoding.
 #[prebindgen]
-pub fn encoding_get_id(e: &Encoding) -> i32 {
-    e.id() as i32
+pub fn encoding_get_id(e: &Encoding) -> u16 {
+    e.id()
 }
 
 /// Create an independent copy of an encoding.
@@ -35,7 +35,7 @@ pub fn encoding_to_string(e: &Encoding) -> String {
 #[derive(Clone, Debug)]
 pub struct EncodingStruct {
     /// Numeric identifier of the encoding.
-    pub id: i32,
+    pub id: u16,
     /// Schema associated with the encoding, when present.
     pub schema: Option<Vec<u8>>,
 }
@@ -70,9 +70,13 @@ pub fn encoding_new_from_string(s: String) -> Encoding {
 /// The schema is raw bytes and is stored verbatim (see [`encoding_get_schema`]),
 /// so this is the exact inverse of [`encoding_get_id`] / [`encoding_get_schema`]:
 /// the pair round-trips losslessly even for a schema that is not valid UTF-8.
+///
+/// The identifier is zenoh's own `EncodingId`, so every value it can hold is a
+/// valid encoding id: there is no out-of-range case to reject and nothing is
+/// narrowed on the way in.
 #[prebindgen]
-pub fn encoding_new_from_id(id: i32, schema: Option<Vec<u8>>) -> Encoding {
-    Encoding::new(id as u16, schema.map(Into::into))
+pub fn encoding_new_from_id(id: u16, schema: Option<Vec<u8>>) -> Encoding {
+    Encoding::new(id, schema.map(Into::into))
 }
 
 /// Create an encoding with the supplied schema.
